1st Attempt at making it Work (and lots of int8_t facepalming later)

+++ packets.h
@@ -3,7 +3,7 @@
 #define POLLRATE 0.1 // 100ms
 
 typedef struct ControlPacket {
-  int8_t pwm[2];
+  int16_t pwm[2];
 } ControlPacket;
 
 typedef struct StatusPacket {
